requirements.txt
absl-py==0.1.12
appnope==0.1.0
astor==0.6.2
beautifulsoup4==4.6.0
bleach==1.5.0
Bottleneck==1.2.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
config==0.3.9
coremltools==0.8
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
Flask==1.0
gast==0.2.0
graphviz==0.8.2
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kaggle==1.1.0
Keras==2.1.5
kiwisolver==1.0.1
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mmdnn==0.1.5
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.5
notebook==5.4.1
numpy==1.14.2
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
psutil==5.4.5
ptyprocess==0.5.2
pydot==1.2.4
Pygments==2.2.0
pyparsing==2.2.0
pyTelegramBotAPI==3.6.2
python-dateutil==2.7.1
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
tensorboard==1.7.0
tensorflow==1.7.0
tensorflow-hub==0.1.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
-e git+https://github.com/tf-coreml/tf-coreml.git@e7e6399ca2c6fe8e980cbf78bcb4532a18679609#egg=tfcoreml
tornado==5.0.1
tqdm==4.23.0
traitlets==4.3.2
urllib3==1.22
uuid==1.30
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.4
xgboost==0.71
